Switch to centos:8 base images
This update change our execution environment to centos:8. The main
change, is the removal of kubernetes-client. This doesn't exist as an
RPM right now in centos.

We also have to enable EPEL to pull in a few packages, needed by
collections. A future update may filter this to only be enabled for the
specific packages we need in EPEL.

Depends-On: https://github.com/ansible/ansible-runner/pull/581
Signed-off-by: Paul Belanger <pabelanger@redhat.com>