Temporarily move away from ansible-builder generated Containerfile
We're trying to get awx-ee running in an ARM environment, but none of the base images used by ansible-builder provide ARM-based images.

While this still utilizes ansible-builder's ability to autodetect collection dependencies, it does not use any of the ansible-runner, python-builder, or python-base base images.

We will move back to ansible-builder generated files at some point in the future.